backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
sentence-transformers==2.2.2
torch==2.2.0+cpu    # see note below
scikit-learn==1.2.2
pdfminer.six==20221105
python-docx==0.8.11
gunicorn==20.1.0
numpy==1.26.0
